refactor+feat: swapchain + image creation, minor refactor of all parts

+++ engine/include/sf_core/application.hpp
@@ -19,8 +19,8 @@ struct ApplicationState {
     sf::PlatformState           platform_state;
     GameInstance*               game_inst;
     Clock                       clock;
-    i16                         width;
-    i16                         height;
+    u32                         width;
+    u32                         height;
     bool                        is_running;
     bool                        is_suspended;
 };
